requirements.txt
flask==3.1.*
gunicorn==23.*
telethon==1.37.*
python-dotenv==1.*
requests==2.*
tqdm==4.*
